completed 90% of the functionality, no UI
This was actually so much fun. managed to get something to run on the computer
I actually did problem solving and math and whatnot here. Gave me back the thrill of problem solving with code after a pretty long time, so far have just been doing html and css xD
Also, got myself familiarized with javascript, i skipped through a lot
of javascript assignments, so had to google quite a bit, but nice i got
comfortable with it already